backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pdfplumber==0.10.3
anthropic==0.18.1
pydantic==2.5.3
python-dotenv==1.0.0
pymupdf==1.24.0
